+++ schema/bool.go
@@ -23,6 +23,10 @@ func (*Bool) Type() ValueType {
 	return TypeBool
 }
 
+func (*Bool) Size() int {
+	return 1
+}
+
 func (dst *Bool) Equal(src CQType) bool {
 	if src == nil {
 		return false