+++ schema/float8_test.go
@@ -39,3 +39,29 @@ func TestFloat8Set(t *testing.T) {
 		}
 	}
 }
+
+func TestFloat8_Size(t *testing.T) {
+	tests := []struct {
+		name string
+		f    Float8
+		want int
+	}{
+		{
+			name: "present",
+			f:    Float8{Float: 1, Status: Present},
+			want: 8,
+		},
+		{
+			name: "null",
+			f:    Float8{Status: Null},
+			want: 8,
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if got := tt.f.Size(); got != tt.want {
+				t.Errorf("Float8.Size() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}